+++ apps/web/pages/invite-user/index.ts
@@ -1,0 +1,3 @@
+import InviteUserListModule from '../../src/Modules/businessUsers/invite.user.list.module';
+
+export default InviteUserListModule;
