+++ apps/web/src/Modules/businessUsers/business.users.list.module.tsx
@@ -5,14 +5,13 @@ import {
 } from '@finnoto/core';
 import { Badge, ConfirmUtil } from '@finnoto/design-system';
 
+import GenericDocumentListingComponent from '../../Components/GenericDocumentListing/genericDocumentListing.component';
 import { GenericDocumentListingProps } from '../../Components/GenericDocumentListing/genericDocumentListing.types';
-import GenericAnimatedTabListing from '../../Components/GenericDocumentListing/genericTabAnimation.component';
-import { openBusinessUserInviteModal } from './business.user.invite.modal';
 
 import { DeleteSvgIcon } from 'assets';
 
 const BusinessUsersListModule = () => {
-    const { removeUser, removeBusinessUser } = useBusinessUserHook({
+    const { removeBusinessUser } = useBusinessUserHook({
         refetch: () => {
             RefetchGenericListing();
         },
@@ -21,7 +20,7 @@ const BusinessUsersListModule = () => {
     const { isOwner, user } = useUserHook();
 
     const props: GenericDocumentListingProps = {
-        name: 'Business Users',
+        name: 'Team Management',
         type: 'business_users',
         table: [
             {
@@ -47,13 +46,6 @@ const BusinessUsersListModule = () => {
             { name: 'Dialling Code', key: 'dialing_code' },
             { name: 'Mobile', key: 'mobile' },
         ],
-        actions: [
-            {
-                name: 'Invite User',
-                type: 'create',
-                action: openBusinessUserInviteModal,
-            },
-        ],
         rowActions: [
             {
                 name: 'Remove',
@@ -78,62 +70,8 @@ const BusinessUsersListModule = () => {
             },
         ],
     };
-    const inviteUserProps: GenericDocumentListingProps = {
-        name: 'Invite Users',
-        type: 'invite_users',
-        table: [
-            { name: 'Email', key: 'email' },
-            { name: 'Invited At', key: 'created_at', type: 'date_time' },
-            { name: 'Accepted At', key: 'accepted_at', type: 'date_time' },
-            { name: 'Rejected At', key: 'rejected_at', type: 'date_time' },
-        ],
-        actions: [
-            {
-                name: 'Invite User',
-                type: 'create',
-                action: openBusinessUserInviteModal,
-            },
-        ],
-        rowActions: [
-            {
-                name: 'Remove',
-                icon: DeleteSvgIcon,
-                type: 'inner',
-                isCancel: true,
-                visible: (data) => {
-                    if (data?.accepted_at || data?.rejected_at) return false;
-                    return true;
-                },
-                action: (data) => {
-                    ConfirmUtil({
-                        isArc: true,
-                        appearance: 'error',
-                        title: `Remove ${data?.name} from the system ?`,
-                        message:
-                            'Are you sure you want to remove this user? This action cannot be undone and will permanently remove the user from the system.',
-                        isReverseAction: true,
-                        onConfirmPress: () => removeUser(data?.id),
-                    });
-                },
-            },
-        ],
-    };
-    return (
-        <GenericAnimatedTabListing
-            tabs_item={[
-                {
-                    tabName: 'Business Users',
-                    tabKey: 'business_user',
-                    listingProps: props,
-                },
-                {
-                    tabName: 'Invited Users',
-                    tabKey: 'invited_users',
-                    listingProps: inviteUserProps,
-                },
-            ]}
-        />
-    );
+
+    return <GenericDocumentListingComponent {...props} />;
 };
 
 export default BusinessUsersListModule;
